daily-digest/requirements.txt
requests>=2.28.0
feedparser>=6.0.0
PyYAML>=6.0
rich>=13.0.0
python-dateutil>=2.8.0
